requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-filter==2.3.0
django-guardian==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
importlib-metadata==1.6.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==1.4.2
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
